Remove duration and data limit.
The QM API exposes time and data limit for jobs but I haven't found a
use for those and only had issues when forgetting to set them high
enough. For now, I've set them to infinity and the only timeout left
is a 1 hour timeout to wait for the results.